requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
numpy==1.23.5
pillow==9.5.0
matplotlib==3.7.1
torch==2.1.2
torchvision==0.16.2
timm==0.9.12
grad-cam==1.4.0
lime==0.2.0.1
scikit-image==0.21.0
opencv-python==4.7.0.72
